pyproject.toml
[project]
name = "distilled-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "datasets>=2.14.0",
    "tqdm>=4.65.0",
    "transformers>=5.2.0",
    "tokenizers>=0.15.0",
    "google-cloud-aiplatform>=1.135.0",
    "setuptools>=82.0.0",
    "pylatexenc>=2.10",
    "ftfy>=6.3.1",
]
